Complete 02-margin-and-padding
First, I added padding to the card to create
borders for the blue sections inside.

All the blue sections have 8px padding from their
content to the borders.

There is a 8px margin between the blue sections
vertically.

The middle blue section has noticeable padding in
the height, so 16px apart.

For the last section, I needed to center the
"and a". To center the button div, I have to use
margin: 0 auto. The only way to align elements
like buttons is to use the margin. The auto helps
a lot. Also because button by default has display
value inline, I have to change it to block.

The rest of the divs such as title, content, and
button container have default values of block, so
I don't need to separate them into blocks.

This was good practice.